+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-build.23] - [2021-05-??]
+### Added
+- It is now possible to change the type of poster that is displayed for media items
+- Item detail pages have a backdrop & poster
+### Changed
+### Deprecated
+### Fixed
+### Removed
+
 ## [v1.0.0-build.22] - [2021-05-10]
 ### Added
 ### Changed